Add template context for AIP-39 variables
The following fields are deprecated:

* execution_date (renamed to schedule_date)
* next_execution_date (renamed to next_schedule_date)
* prev_execution_date (see below)
* prev_execution_date_success (renamed to prev_schedule_date_success)
* prev_ds
* prev_ds_nodash
* tomorrow_ds
* tomorrow_ds_nodash
* yesterday_ds
* yesterday_ds_nodash

A new field prev_schedule_date is added to replacement the functionality
provided by prev_execution_date, but with a different implementation.
The old 'execution' one is calculated in abstract by substracting an
interval from execution_date/schedule_date, but the new 'schedule' one
is based on the *actual* previous DagRun recorded in the database. The
new behavior matches prev_execution_date_success and is considered a
better implementation, but introduces incompatibility in certain edge
cases, such as when a DAG is first run (ever, or after a pause/unpause).

The following new fields are added for the new data interval concept
introduced in AIP-39:

* data_interval_start
* data_interval_end

This commit also includes a minor refactor on how a TaskInstance without
an associated DagRun is handled. This is impossible in practice, but is
depended on by many tests in Airflow's main test suite.